requirements.txt
secure-smtplib
python-decouple==3.3